+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Improve font loading on skins.
 - Add new "Share on Mastodon" button.
 - Improve Universal skin.
+- Remove Clipboard.js dependency. Chulapa now uses a custom script.
 
 
 ## v1.0.1 - 2022-11-25